some warning corrections under arduinopio and esp8266

+++ TM1637TinyDisplay.cpp
@@ -432,7 +432,7 @@ void TM1637TinyDisplay::showString(const char s[], uint8_t length, uint8_t pos, 
 
   // Basic Display
   if (strlen(s) <= MAXDIGITS) {
-    for (int x = 0; x < strlen(s); x++) {
+    for (size_t x = 0; x < strlen(s); x++) {
       digits[x] = encodeASCII(s[x]);
     }
     if(dots != 0) {
@@ -483,7 +483,7 @@ void TM1637TinyDisplay::showString_P(const char s[], uint8_t length, uint8_t pos
 
   // Basic Display
   if (strlen_P(s) <= MAXDIGITS) {
-    for (int x = 0; x < strlen_P(s); x++) {
+    for (size_t x = 0; x < strlen_P(s); x++) {
       digits[x] = encodeASCII(pgm_read_byte(&s[x]));
     }
     if(dots != 0) {
@@ -614,7 +614,7 @@ bool TM1637TinyDisplay::Animate(bool loop)
         case 3: // PROGMEM text scroll running
             for (int x = 0; x < MAXDIGITS; x++) {
                 int offset = frame_num - MAXDIGITS + x;
-                if (offset >= 0 && offset < m_animation_frames - (2 * MAXDIGITS)) {
+                if (offset >= 0 && offset < (int)(m_animation_frames - (2 * MAXDIGITS))) {
                     digits[x] = encodeASCII(pgm_read_byte(&m_animation_string[offset]));
                 } else {
                     digits[x] = 0;
@@ -625,7 +625,7 @@ bool TM1637TinyDisplay::Animate(bool loop)
         case 4: // SRAM text scroll running
             for (int x = 0; x < MAXDIGITS; x++) {
                 int offset = frame_num - MAXDIGITS + x;
-                if (offset >= 0 && offset < m_animation_frames - (2 * MAXDIGITS)) {
+                if (offset >= 0 && offset < (int)(m_animation_frames - (2 * MAXDIGITS))) {
                     digits[x] = encodeASCII(m_animation_string[offset]);
                 } else {
                     digits[x] = 0;
